pyproject.toml
[tool.poetry]
name = "grave-digger"
version = "0.1.0"
description = "El Grave Digger App"
authors = ["Bruno Gazineu <brunosgazineu@hotmail.com>"]

[tool.poetry.dependencies]
python = "3.11.4"
gunicorn = "21.2.0"
psycopg2-binary = "~2.9.5"
django = "4.2.7"
django-extensions = "^3.2.3"
django-environ = "^0.9.0"
django-split-settings = "1.2.0"
djangorestframework = "3.14.0"
django-cors-headers = "4.3.1"